# public/openapi.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "REVREBEL Modular SEO & Hotel Commercial Audit API",
        "version": "1.0.0",
        "description": "Enterprise website evaluation pipeline optimizing technical SEO, hospitality telemetry schema mapping, and performance metrics tracking."
    },
    "servers": [
        {
            "url": "https://seo-api.revrebel.io",
            "description": "Production Environment"
        }
    ],
    "paths": {
        "/health": {
            "get": {
                "summary": "Health Status Verification",
                "responses": {
                    "200": {
                        "description": "Service online confirmation tracker",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "UP"
                                        },
                                        "timestamp": {
                                            "type": "string",
                                            "format": "date-time"
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/audit": {
            "post": {
                "operationId": "auditSeoPage",
                "summary": "Execute Commercial Site Diagnostic Evaluation",
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "required": [
                                    "url"
                                ],
                                "properties": {
                                    "url": {
                                        "type": "string",
                                        "format": "uri",
                                        "example": "https://example.com"
                                    },
                                    "renderMode": {
                                        "type": "string",
                                        "enum": [
                                            "static",
                                            "browser"
                                        ],
                                        "default": "static"
                                    },
                                    "includePerformance": {
                                        "type": "boolean",
                                        "default": false
                                    },
                                    "includeAccessibility": {
                                        "type": "boolean",
                                        "default": false
                                    },
                                    "viewport": {
                                        "type": "string",
                                        "enum": ["desktop", "tablet", "mobile"],
                                        "default": "desktop",
                                        "description": "Viewport preset for browser rendering mode."
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Complete Consolidated Structural Audit Payload Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "auditId": {
                                            "type": "string",
                                            "format": "uuid"
                                        },
                                        "status": {
                                            "type": "string",
                                            "example": "completed"
                                        },
                                        "result": {
                                            "type": "object"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Security layer token authentication violation"
                    },
                    "400": {
                        "description": "SSRF or invalid URL specification rejection payload"
                    }
                }
            }
        },
        "/api/audit/{auditId}": {
            "get": {
                "operationId": "getAuditRun",
                "summary": "Retrieve a specific audit run by its ID",
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "parameters": [
                    {
                        "name": "auditId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "The requested audit run",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Audit run not found"
                    }
                }
            }
        },
        "/api/audits": {
            "get": {
                "operationId": "listAuditRuns",
                "summary": "List recent audit runs",
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "parameters": [
                    {
                        "name": "domain",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "schema": {
                            "type": "integer",
                            "default": 10
                        }
                    },
                    {
                        "name": "offset",
                        "in": "query",
                        "schema": {
                            "type": "integer",
                            "default": 0
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "A list of recent audit runs",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/url-scan/import": {
            "post": {
                "operationId": "importUrlScan",
                "summary": "Import a completed raw URL scanner payload.",
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "requestBody": {
                    "required": true,
                    "description": "Accepts either a wrapped payload structure specifying the provider, or the raw complete scanner JSON file object directly.",
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Imported scanner result",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "scanId": {
                                            "type": "string",
                                            "format": "uuid"
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/url-scan": {
            "post": {
                "operationId": "runUrlScan",
                "summary": "Run a new URL scanner request.",
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "required": [
                                    "url"
                                ],
                                "properties": {
                                    "url": {
                                        "type": "string",
                                        "format": "uri"
                                    },
                                    "provider": {
                                        "type": "string",
                                        "default": "cloudflare"
                                    },
                                    "waitForResult": {
                                        "type": "boolean",
                                        "default": false
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "202": {
                        "description": "Scanner submission accepted"
                    },
                    "200": {
                        "description": "Scanner result returned synchronously"
                    }
                }
            }
        },
        "/api/url-scan/{scanId}": {
            "get": {
                "operationId": "getUrlScan",
                "summary": "Retrieve a stored URL scanner result.",
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "parameters": [
                    {
                        "name": "scanId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    },
                    {
                        "name": "includeRaw",
                        "in": "query",
                        "schema": {
                            "type": "boolean",
                            "default": false
                        }
                    },
                    {
                        "name": "includeRequests",
                        "in": "query",
                        "schema": {
                            "type": "boolean",
                            "default": false
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "The requested URL scan"
                    },
                    "404": {
                        "description": "URL scan not found"
                    }
                }
            }
        },
        "/api/url-scans": {
            "get": {
                "operationId": "listUrlScans",
                "summary": "List stored URL scanner results.",
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "parameters": [
                    {
                        "name": "domain",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "apexDomain",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "sourceProvider",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "schema": {
                            "type": "integer",
                            "default": 20
                        }
                    },
                    {
                        "name": "offset",
                        "in": "query",
                        "schema": {
                            "type": "integer",
                            "default": 0
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "A list of URL scans"
                    }
                }
            }
        },
        "/api/url-scan/{scanId}/refresh": {
            "post": {
                "operationId": "refreshUrlScanResult",
                "summary": "Refresh a provider-backed URL scanner result.",
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "parameters": [
                    {
                        "name": "scanId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Refreshed URL scan result"
                    },
                    "400": {
                        "description": "Scan cannot be refreshed"
                    },
                    "404": {
                        "description": "URL scan not found"
                    }
                }
            }
        }
    },
    "components": {
        "securitySchemes": {
            "ApiKeyAuth": {
                "type": "apiKey",
                "in": "header",
                "name": "x-api-key"
            }
        }
    }
}